fine turned color, optimized memory usage

+++ Main/Led.cpp
@@ -24,7 +24,7 @@ Led::Led(uint8_t index)
     _busy = false;
 }
 
-void Led::newValue(uint16_t r, uint16_t g, uint16_t b)
+void Led::newValue(uint8_t r, uint8_t g, uint8_t b)
 {
     new_r = r;
     new_g = g;
